Laptop-Price-Predictor
I Developed a machine learning model that can accurately predict the price of a laptop based on its various features and specifications. The goal is to develop a model that can accurately estimate the cost of a laptop given its attributes such as processor type, RAM size, storage capacity, graphics card, operating system (OS), brand, and other relevant factors. The dataset for training and testing the model will include historical data on laptops with corresponding prices. The model should be able to handle a wide range of laptop configurations and accurately predict prices for both new and used laptops.